requirements.txt
streamlit>=1.44,<2.0
pandas>=2.2,<3.0
PyYAML>=6.0,<7.0
